# Fix undo history