--- a/src/selection.rs
+++ b/src/selection.rs
@@ -1,13 +1,12 @@
 /// Everything related to the caret position and DOM selection ranges.
 
 use wasm_bindgen::JsCast;
-use web_sys::{Node, Range, Text};
+use web_sys::{Node, Selection, Range, Text};
 
 /// A position relative to a node.
 #[derive(Debug)]
 pub enum Position<'a> {
     /// Caret position is before the selected node.
-    #[allow(dead_code)]  // Needed in tests
     Before(&'a Node),
 
     /// Caret position is after the selected node.
@@ -70,6 +69,13 @@ pub fn set_selection_range(start: &Position, end: Option<&Position>) -> Option<R
         None => range.collapse_with_to_start(true),
     }
 
+    activate_selection_range(&selection, &range);
+    Some(range)
+}
+
+/// Activate the specified selection range in the DOM. Remove all previous
+/// ranges.
+pub fn activate_selection_range(selection: &Selection, range: &Range) {
     // Note: In theory we don't need to re-add the range to the document if
     //       it's already there. Unfortunately, Safari is not spec-compliant
     //       and returns a copy of the range instead of a reference when using
@@ -79,8 +85,6 @@ pub fn set_selection_range(start: &Position, end: Option<&Position>) -> Option<R
     //       See https://bugs.webkit.org/show_bug.cgi?id=145212
     selection.remove_all_ranges().unwrap();
     selection.add_range(&range).expect("Could not add range");
-
-    Some(range)
 }
 
 #[cfg(test)]