Unfortunately DOM manipulations that don't use `document.execCommand` aren't added to the undo stack. So we have to resort to inserting HTML strings using the `insertHTML` command.

This increases the benchmark time a bit:

- Insert text: 260 us -> 320 us
- Insert image: 390 us -> 910 us

Still below 1 ms though, so all is fine. The code is considerably more ugly though :slightly_frowning_face: 

Fixes #42.